backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.0
reportlab==4.2.0
slowapi==0.1.9
scikit-learn==1.6.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
python-multipart==0.0.9
bcrypt==4.1.3
